Map memory consumption optimizations.
64bit warning fixes.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@655 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 780e67b09257648104958fb12dae547ca90be970

+++ Source/AngelScript/reflection.cpp
@@ -226,7 +226,7 @@ ScriptString* ScriptType::GetPropertyDeclaration( uint index, bool include_names
 uint ScriptType::GetEnumLength() const
 {
     asCObjectType* objTypeExt = (asCObjectType*) ObjType;
-    return objTypeExt->enumValues.GetLength();
+    return (uint) objTypeExt->enumValues.GetLength();
 }
 
 ScriptArray* ScriptType::GetEnumNames() const
